Make it possible to specify the base commit for each ref in linear rebase (#48)
Previously, linear rebase assumed that the stack is already being
linear. This works if it's just rebasing the entire stack to the latest
main, but it's not enough if we want to restack a part of the stack.

This changes the linear rebase argument to take a base commit for each
ref so that it can be used for other cases such as restacking a
partially merged stack.