Use the right resolver (#57)
The linear rebase implementation does a merge operation twice. Once to
discover which blobs need to be fetched in order to resolve the
conflicts, and once to actually perform the merge after fetching the
blobs. The first merge uses a "collector" resolver that just collects
the missing blobs, while the second merge should use the actual resolver
that knows how to fetch the missing blobs.

This change fixes a bug where the second merge was incorrectly using
the collector resolver instead of the actual resolver.