Add get-branching-point
This is a bit more efficient version for finding a branching point for
feature branches. The limitation is that the feature branch should have
a linear history (= no merge commit).

The command expects that the branching point is found relatively at a
shallow position. It exponentially deepens the depth for exploration.