Add an option to abort the backport
This is a breaking change. Currently, if there's a conflict during a
backport, it automatically aborts the operation. With this change,
niche-git will continue the backport process even if conflicts are
detected. The conflicting content will be marked if it's a text file, or
other cases, there will be a .rej file created etc.. The caller can
still get the same behavior with the new `AbortOnConflict` option.